null pointer exception fix: Pull Request to solve a nullPointerException produced when recreating an activity with fragments after it has been killed by the OS.
You can easily reproduce the problem if you check
Settings -> Developer Options -> Don't keep activities.
The activity with fragments will be killed and when you get back you will
get this NullPointerException.

To solve this, I have inverted the order of a couple of sentences in the
onCreate method of DaggerActivity and DaggerFragmentActivity